pyproject.toml
[tool.poetry]
name = "dank-mids"
version = "4.20.131"
description = "Multicall batching middleware for asynchronous scripts using web3.py"
authors = ["BobTheBuidler <bobthebuidlerdefi@gmail.com>"]
homepage = "https://github.com/BobTheBuidler/dank_mids"
repository = "https://github.com/BobTheBuidler/dank_mids"
documentation = "https://BobTheBuidler.github.io/dank_mids"
license = "MIT"
readme = "README.md"
include = [
    { path = "5eab82354e25a032f547__mypyc.cpython-38-x86_64-linux-gnu.so", format = ["sdist", "wheel"] },
    { path = "5eab82354e25a032f547__mypyc.cpython-38-darwin.so", format = ["sdist", "wheel"] },
    { path = "5eab82354e25a032f547__mypyc.cp38-win_amd64.pyd", format = ["sdist", "wheel"] },
    { path = "c6c6c8f11e2c85534eef__mypyc.cpython-39-x86_64-linux-gnu.so", format = ["sdist", "wheel"] },
    { path = "c6c6c8f11e2c85534eef__mypyc.cpython-39-darwin.so", format = ["sdist", "wheel"] },
    { path = "c6c6c8f11e2c85534eef__mypyc.cp39-win_amd64.pyd", format = ["sdist", "wheel"] },
    { path = "c6c6c8f11e2c85534eef__mypyc.cpython-310-x86_64-linux-gnu.so", format = ["sdist", "wheel"] },
    { path = "c6c6c8f11e2c85534eef__mypyc.cpython-310-darwin.so", format = ["sdist", "wheel"] },
    { path = "c6c6c8f11e2c85534eef__mypyc.cp310-win_amd64.pyd", format = ["sdist", "wheel"] },
    { path = "c6c6c8f11e2c85534eef__mypyc.cpython-311-x86_64-linux-gnu.so", format = ["sdist", "wheel"] },
    { path = "c6c6c8f11e2c85534eef__mypyc.cpython-311-darwin.so", format = ["sdist", "wheel"] },
    { path = "c6c6c8f11e2c85534eef__mypyc.cp311-win_amd64.pyd", format = ["sdist", "wheel"] },
    { path = "c6c6c8f11e2c85534eef__mypyc.cpython-312-x86_64-linux-gnu.so", format = ["sdist", "wheel"] },
    { path = "c6c6c8f11e2c85534eef__mypyc.cpython-312-darwin.so", format = ["sdist", "wheel"] },
    { path = "c6c6c8f11e2c85534eef__mypyc.cp312-win_amd64.pyd", format = ["sdist", "wheel"] },
]

[tool.poetry.dependencies]
python = ">=3.8,<3.13"
aiofiles = "*"
aiolimiter = ">=1.2,<1.3"
cchecksum = ">=0.0.3,<1"
eth-retry = ">=0.3.1,<1"
evmspec = ">=0.2.8,<1"
ez-a-sync = ">=0.28,<1"
multicall = ">=0.6.2,<1"
typed-envs = ">=0.0.8,<0.1"
web3 = ">=5.27,!=5.29.*,!=5.30.*,!=5.31.1,!=5.31.2,<8"

[tool.poetry.group.dev.dependencies]
eth-brownie = [
    {version = ">=1.16,<1.20", python = ">=3.8,<3.10"},
    {version = ">=1.20.3", python = ">=3.10,<4"}
]
pytest = "*"
pytest-asyncio-cooperative = "*"
pytest-cov = "*"
pytest-sugar = "*"
# need this to resolve brownie locally, didn't before
importlib-resources = ">=6.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
exclude = ["build/","cache/","env/","tests/"]
ignore_missing_imports = true
check_untyped_defs = true
disable_error_code = ["return"]

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | cache
  | docs
  | dist
  | env
)/
'''
